front: importtrainlist: move payload posting to helper file
